cambio en api
 cambio en login de metodo ObtenerPorUsuario al de AccesoUsuario para se valide usuario y contraseña y no solo usuario, ademas del endpoint ObtenAperturaAbierta que luego servira para cosas como corte de caja, validar inicio de venta,

+++ Voalaft.Data/Entidades/RegAperturaCaja.cs
@@ -17,7 +17,7 @@ namespace Voalaft.Data.Entidades
         public decimal DotacionInicial { get; set; }
         public int IDEmpleado { get; set; }
         public int? IDUsuarioAutoriza { get; set; }
-        public byte Estatus { get; set; }
+        public int Estatus { get; set; }
         public bool Activo { get; set; }
         public string? Usuario { get; set; }
         public string? Maquina { get; set; }
